Move things from dbt.contracts.util to dbt.artifacts.base
